fix(KUI-1140): moving semesterDropdown to component-shared
